HACK mgr: pass through cluster log as notify() calls
This is hacky because we're sending log messages
through as strings as the second argument
to notify, which is generally meant to be
a version indicator not the data.

We should either have a separate notify-like
call that has explicit args for the fields
in a log message, or we should build a python
structure to pass through into notify() so
that we're not trashing the semantics of
the log entry.

This also doesn't address the need for plugins
to be able to declare what level of logging
they want to see to avoid getting lots of
unwanted callbacks.

Signed-off-by: John Spray <john.spray@redhat.com>